Reduce memory copy in realtime-only encoding mode
When in realtime encoding more without deblocking, it was shown in
profiling that memory copy consumes most of encoder's time. These
operations mostly take place in vp8_build_inter16x16_predictors_mb.

Many of these memory copies are due to frame reconstruction for
ZEROMV either because there's no motion in the frame or motion is
disabled by active map.

One observation is that the encoder rotate buffers for encoding
the latest frame and most of the time it flips between two buffers.
We can take advantage of this behavior such that if we detect that
copy happens from frame A->B and there is a copy from frame B->A
then we can skip the second copy because source and destionation
has the same content.

This patch introduces a new data structure mb_copy_ref to keep
tracking of this information.

On frames with large area disabled by active map this change
reduces encoding time by ~20%.

Change-Id: I0bb2ba1a9acac4b6d952c4cf3fdc63eeea81f96b