Fixes necessary for input partitions interface to match the RTP profile
These changes fixes a glitch between the RTP profile and the input partitions
interface. Since there's no way for the user to know the actual number of
partitions, the decoder have to read the multi_token_paritition bits also
when input partitions mode is enabled.

Included are also a couple of fixes for issues with independent partitions
and uninitialized memory reads.

Change-Id: I6f93b15287d291169ed681898ed3fbcc5dc81837