WIP: revised loopfilter
The data processed by the loopfilter overlaps. At the block level, this
results in some redundant transforms. Grouping the filtering allows for
a single 16x16 transpose (actually two) instead of three 16x8 transposes
(actually six)

This implementation is linux (and I think mac) x86_64 only. Comments
requested specifically on the transpose in the vertical filter.

Improvements are obviously material dependant, but it seems to be ~%1 in
tests here

Change-Id: I467b7ec3655be98fb5f1a94b5d145e5e5a660007