[Experiment] Skip loopfiltering if a mb and its surrounding friends are static for a while
If a mb, the one to the top and the one to the left are static for 2 consecutive frames
then skip loopfiltering. The rationale is that they are filtered already in the last
frame so doing an extra filtering doesn't help.

I have not taken into account sharpening so doing this may have an adverse effect.

Change-Id: I2d2047485d899489dc726526e8ce8498dff64d2a