Update patch set 1
Patch Set 1:

Is it possible that we put this behind a feature bit? So in the bitstream we do this only if user select this feature, this way we can ensure both sides have the same version. This is more intended for realtime encoding which would be easier to confirm that both sides do loop filtering the same way.

Patch-set: 1