Added reuse partition for multiple encodes
This commit enables to reuse the first partition data from
the first encode for the subsequent encodes. Necessary data
such as quantization, loopfilter etc. are overwritten in
the first partition for the other encodes. The flag
REUSE_PARTITION in blockd.h needs to be enabled for this
feature. Initial results show savings of around 25% in
speed and 10% in datarate. The quality loss is around
0.1 to 0.2dB for close datarates between encodes.

Also fixed a bug related to tokenization when
mb_no_coeff_skip is disabled in alloccommon.c

Change-Id: I365cfb6ac636f19c545f682e3aeac185253abaef