Update patch set 3
Patch Set 3: (3 inline comments)

I might not have explained chunks well enough. Chunk #n is defined as containing at least partition #n, but can also contains several partitions, or parts of the next partition. 
 So we can for example have: 
[0]: part0 + part1 
[1]: empty 
[2]: part2 + half of part3 
[3]: other half of part3 
[4]: part4

We need support for this because there's a way of packetizing the payload into RTP packets in an aggregated way, so that a single packet can contain several partitions (and parts of a partition).

I was considering renaming the partition pointers to chunk[] and chunk_sizes[], to clarify the behavior. What do you think?

Patch-set: 3
Label: Code-Review=0
Label: Verified=0