Update patch set 3
Patch Set 3:

ok.. I think I understand. I'd want to see this implemented so that normal case (old api) is handled as a single "chunk".

So I think I understand it as so:
 Chunks are contiguous.
 Chunks can contain multiple partitions
 The last partition in the chunk is potentially incomplete
 When you start parsing a chunk you know which partition the first byte belongs to.

So there's probably some renaming in order, ie make it clear in the api and through the vpx_dx_iface.c shim that you're talking about chunks, and this routine that we're reviewing is the bit that turns chunks into partitions.

Nit: I wonder if we should use fragment vice chunk, since that's I think consistent with the term used in the RTP spec.

Patch-set: 3