Update patch set 3
Patch Set 3:

wasn't reading that clearly ... there is an 'else' hidden in there.

It looks like we always prefer the no_extend version for neon. If it's built with neon and rtcd, it will only use the ptr function if we *don't* actually have neon. So I'm in favor of removing the yonly neon function.

Patch-set: 3