reduce copy during loop_filter_pick
DO NOT SUBMIT: still needs cleanup

only y is filtered. also, during filtering, the borders are not touched.
it skips MB filtering for row/col == 0

because the C yonly copy is only used for this function, we don't need
to extend the borders. because the neon version already recognized this,
we can remove the neon full Y copy function (not actually removed yet)

names still need to be changed to clean things up a little

Change-Id: Ic549adfca70fc6e0b55f8aab0efe81f0afac89f9