added code to clear 2nd order block when appropriate
It is discovered that in rare situations the 2nd order block may
produce a few small magnitude coefficients that has no effect on
reconstruction. This commit added code to detect such odd  cases
and reset the 2nd order block to all 0.

Test on derf set shows a consistent gain of .12%-.14% on all
metrics.

Change-Id: I718e19cf59a4fc2462cb7070832759beb9f7e7dd