added code to clear 2nd order block when appropriate
It is discovered that in rare situations the 2nd order block may
produce a few small magnitude coefficients that has no effect on
reconstruction. The situations are a combination of low quantizer
values (high quality) and low energy in residual signals (content
dependent). This commit added code to detect such cases and reset
the 2nd order block to all 0.

This commit uses code to do all-zero-check on idct result buffer,
potential speed optimizations are to do the check as part of the
idct and only do the check when effective 2nd order quantizer is
small (lower than 64).

Test on derf set shows a consistent gain of .12%-.14% on all
metrics.

Change-Id: I718e19cf59a4fc2462cb7070832759beb9f7e7dd