added code to clear 2nd order block when appropriate
It is discovered that in rare situations the 2nd order block may
produce a few small magnitude coefficients that has no effect on
reconstruction. The situations are a combination of low quantizer
values (high quality) and low energy in residual signals (content
dependent). This commit added code to detect such cases and reset
the 2nd order block to all 0.

Patch 1 to 4 used code to do all-zero-check on idct result buffer,
and tests on derf set showed a consistent gain of .12%-.14% on all
metrics. But due to a recent change Ie31d90b, the idct result buffer
is not longer populated. So patch 5 uses an alternative method to
detect the situations. Tests on derf set now shows a consistent
quality gain of .16%-.20%. Most of the gains are coming from high
quality range, as the change has no effect on low quality range.

Change-Id: I718e19cf59a4fc2462cb7070832759beb9f7e7dd