Update patch set 4
Patch Set 4:

I saw scott's recent change that had idct/recon combined, so we need take a closer look to see if the diff buffers are still being populated with the idct results. 

On the other hand, this means an optimization opportunity for recon, i.e. if the results of IDCTs are all zero, recon should be a copy instead, no add and saturation needed

Patch-set: 4
Label: Code-Review=0
Label: Verified=0