DO NOT SUBMIT
elf SSE4 implementation of 4x4 intra predictions

TODO: coff, mach-o

Switch statement implementation is ... rudimentary.

In the majority of cases, all the Left values and the low Above [0-3]
values are required (5/10). Of the remaining cases, two require all of
Above [0-7], and one each require top_left Above[0-4],
Left[0-3] top_left, and Left[0-3]

So for some added complexity, we could check anything that doesn't
require Left[] (because it's the biggest pain to get) and jump directly
to them, where they would require only one load to get their values. But
only 3/10 fall into that case.

The individual functions need to be gone over as well. The later ones
are definitely sub-optimal and need a close look.

.asm file does not conform to any kind of style yet

Now with 32bit elf and without xmm7

Change-Id: I4eb3b900c63424e654bb945d8bb51a9ac445646f