Update patch set 1
Patch Set 1:

Looked into this some more with John and the correct fix is to change build/make/configure.sh around line 904:

 add_asflags -f win${bits}

Apparently win64 and x64 are both acceptable to yasm. VS works because it doesn't pass our asflags to yasm and it uses x64.

So the correct thing to do is add '-f win32' or '-f x64', depending on the target. If you'd like to fix it there I will take the patch (provided you don't add the target to configure until most things appear to be working)

Patch-set: 1