avoid resetting framerate during vpx_codec_enc_config_set()
The calculated frame_rate is a state variable in the codec, and shouldn't
be maintained in the configuration struct. Move it to the main part of cpi
so that it isn't clobbered when the configuration struct is updated. The
initial framerate estimate is moved from the vp8_cx_iface.c wrapper into
the body of init_config() in onyx_if.c, so that it is only called once and
not reset on every call to vp8_change_config().

Change-Id: I8d9a3d1283330d1ee297d07e9d78d1f2875f2465