Remove unnecessary ternary constructs
The code had a number of constructs like (condition)?1:0,
which is redundant with C's semantics. In the cases where a boolean
operator was used in the condition, simply remove the ternary part.
In the cases where the desired result was an exprsession evaluated
in boolean context, wrap the condition as !!(condition).

Change-Id: Icb2372defa3783cf31857d90b2630d06b2c7e1be