Base the ARM subset configuration on target isa
Previously --target=armv7-... --disable-armv7 would
disable armv6 and armv5te because the configuration
of the subsets was based on the highest enabled
and in this case none is enabled because the
highest has been explicitly disabled.
This fix bases the enabling of the subsets on the
target isa, which is derived from the --target
option and not on the already enabled sets.

Change-Id: Ic2978c6a64775bce62b56ca84656b4afbd7296ab