RFC: Reorganize MFQE loops
Break MFQE code into it's own file.

There is a failed experiment at the bottom of the file to combine the
variance and SAD. I saw no difference between the existing assembly and
the new code. To be removed.

Since we're keeping variance and SAD they need to be moved from encoder
to common.

Diminishing returns for 4x4 and 2x2 filter_by_weight. Still
investigating whether they are worthwhile.

Change-Id: I4b8de6b5cdff5413037de5b6c42f437033ee55bf