Update XCode SDK search paths
Newer XCodes have moved the SDK path from /Developer/SDKs to a
resource of the XCode.app itself. This patch loops through SDK
versions and root paths to try to find an SDK it can use, and exits
with an error otherwise. This is still somewhat fragile, as it assumes
XCode is installed in the default location. Might be a way to
introspect via gcc to extract the actual path to use?

Change-Id: I4f5d7e35175d0dea84faaa6bfb52a0153c72f84b