vpxenc: support scaling prior to encoding
Scales the input of the encoder using libyuv's "box filter". Each stream
may have a different width and height specified. If the width (or
height) parameter is missing (or is explicitly set to 0) then the value
will be calculated based on the specified height (or width) and the
input file's dimensions, preserving its aspect ratio. Leaving the height
unspecified behaves similarly.

Note: This functionality still does not take advantage of the
accelerated multi-resolution encoder support with
CONFIG_MULTI_RES_ENCODING.

Change-Id: Ic7026810b13be030826be80dc6f7fc4aaf0c35d0