MFQE: add chroma check to threshold
In cases where you have a flat background occluded by a moving object
of similar luminosity in the foreground, it was likely that the foreground
blocks would persist for a few frames after the background is uncovered.
This is particularly noticable when the object has a different color
than the background, so add the chroma planes in as an additional check.

This fixes issue #392, which includes a test clip to reproduce the
issue.

BUG=392

Change-Id: I2bd7b2b0e25e912dcac342e5ad6e8914f5afd302