Update patch set 1
Patch Set 1:

I think it looks fine, but could you update the commit message with a bit more information.  I thought that it was working previously.  Did an update to the ndk break it?  Did it never work on a clean system?  Is this a win/linux/mac only issue?  Or was this never correctly implemented when the rtcd system was changed over?

Patch-set: 1
Label: Code-Review=0
Label: Verified=0