Update patch set 1
Patch Set 1: Looks good to me, approved

Might be good to wrap this into a function in a future commit, so it's just a one-liner at the point where the sync needs to happen. There might be an argument to be made for putting the calls into the vp8_cx_iface.c wrapper, so that it's easy to enforce that the sync happens before doing *anything*.

Will merge this as-is since it's in the spirit of the current code, just some things to think about especially if we ever get to do a proper overhaul of the multithread code.

Patch-set: 1
Reviewer: Gerrit User 1000856 <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: SUBM=+1
Label: Verified=0