Update patch set 4
Patch Set 4:

This has a couple of separate changes in it now, could have split on the g++ to CXX changes. Also not sure that it's kind to force a c++ compiler install just to build the lib...should the unit tests be enabled by default, but still be allowed to be disabled?

Patch-set: 4
Label: Code-Review=0
Label: Verified=0