Reset Q for key frame when spatial resizing occurs.
The logic for spatial resizing is done after the Q is selected for the frame.
This causes a problem that the Q we select for the (resized) key frame may
be based on a different resolution than the frame we will encode.

This fix is to ensure that, when resize is on, the selected Q is still
based on the resolution of the frame to be encoded.

Change-Id: Ie15b799fca695a5abef4439d8b1fd844796a832d