Reset Q for key frame when spatial resizing occurs.
The logic for spatial resizing is done after the Q is selected for the frame.
This causes a problem that the Q we select for the (resized) key frame may
be based on a different resolution than the frame we will encode.

This fix is to ensure that, when resize is on, the selected Q is still
based on the resolution of the frame to be encoded.

Change-Id: Ia49a9eac5f64e48d1c00dfc7ed4ce26fe84d3fa1