prototype opaque reference counted buffer allocator
This patch is for discussion only at this point, as it doesn't
implement the described functionality.

The idea here is that if you store the reference counter in a way
that's easy to derive from the object's pointer, it can be fast enough
that it can be used as a lightweight synchronization primative, a'la
triple buffering or read-copy-update. This attempts a generic interface
rather than doing something specific to framebuffers, since I think
there's other data that this approach could be valuable for. The idea
is to be thin and fast rather than presenting a vtable to anything
you might want to reference count.

It'd be fairly easy to expose the allocation poritions of this API to
the application, and there was some discussion[1] along these lines
back in Feb. 2011. This doesn't do anything to solve or even improve
the case of direct rendering, but I seem to recall there were some
IPC cases that this would be desirable for.

[1]: http://comments.gmane.org/gmane.comp.multimedia.webm.devel/1402

Change-Id: Ia7d3c98fe2dc30545fbd929a44b28db70646dab5