Clean Android build defaults
Disable unit-tests. The logging in GTest would need to be adjusted.

soft-enable runtime cpu detection. Allows disabling during configure.

Restructure ARM cpu detection. Move the no-op case to the top so we
exit early if there is no RTCD.

Change #if defined(HAVE_*) to #if HAVE_* because we only need to check
for features that the library was actually built with. This should have
been harmless, as disabled feature sets wouldn't have any features to
call.

Change-Id: Iea21aa42ce5f049c53ca0376d25bcd0f36f38284