Update patch set 1
Patch Set 1: (3 inline comments)

I added the new dependencies to the existing README. I'll do a larger "this is how you run the tests" section as part of a later commit, explaining it will be a good excuse to work out some of the wrinkles in the process, there are some clunky things right now.

Not sure what the best place is to put stuff aimed at developers.. Some projects have a HACKING file, maybe we could do something like that. Document coding conventions and that sort of thing in there too. Use gerrit to vote on changes. That's a bigger thing than either of these changes.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0