Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Verified that it builds (though not tested with all possible options), and encode / decode done on a few files.

There are a number of slightly messy outcomes but I think this is inevitable and we should just address them as and when we come across them. Waiting too long on submitting will just hold up other work or make for messy merges.

Patch-set: 1