Remove the resetting of ambient QP for temporal layers.
The ambient QP was reset to worst_allowed for every frame
when temporal layers is on. This change removes this reset,
so ambient QP is what it should be (local average).

Also, as this affects the target size for forced key frames
(it will actually lower the size somewhat), we increased the
inital boost factor to compensate.

Change-Id: I3fb8253445ee192034980d5a55b6fe5a947f2b65